requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-admin-generator==2.2.1
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-phonenumber-field==7.0.1
django-utils2==2.12.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.2
Pillow==9.3.0
PyJWT==2.6.0
python-utils==3.4.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
razorpay==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
whitenoise==6.2.0
